+++ Tests/MASShortcutTests/MASShortcutBinderTests.m
@@ -1,3 +1,9 @@
+#import <Cocoa/Cocoa.h>
+#import <XCTest/XCTest.h>
+#import "MASShortcut.h"
+#import "MASShortcutBinder.h"
+#import "MASDictionaryTransformer.h"
+
 static NSString *const SampleDefaultsKey = @"sampleShortcut";
 
 @interface MASShortcutBinderTests : XCTestCase
@@ -29,7 +35,9 @@ static NSString *const SampleDefaultsKey = @"sampleShortcut";
 - (void) testInitialValueReading
 {
     MASShortcut *shortcut = [MASShortcut shortcutWithKeyCode:1 modifierFlags:1];
-    [_defaults setObject:[NSKeyedArchiver archivedDataWithRootObject:shortcut] forKey:SampleDefaultsKey];
+    NSError *error;
+    [_defaults setObject:[NSKeyedArchiver archivedDataWithRootObject:shortcut requiringSecureCoding:YES error:&error] forKey:SampleDefaultsKey];
+    XCTAssertNil(error);
     [_binder bindShortcutWithDefaultsKey:SampleDefaultsKey toAction:^{}];
     XCTAssertTrue([_monitor isShortcutRegistered:shortcut],
         @"Pass the initial shortcut from defaults to shortcut monitor.");
@@ -39,7 +47,9 @@ static NSString *const SampleDefaultsKey = @"sampleShortcut";
 {
     MASShortcut *shortcut = [MASShortcut shortcutWithKeyCode:1 modifierFlags:1];
     [_binder bindShortcutWithDefaultsKey:SampleDefaultsKey toAction:^{}];
-    [_defaults setObject:[NSKeyedArchiver archivedDataWithRootObject:shortcut] forKey:SampleDefaultsKey];
+    NSError *error;
+    [_defaults setObject:[NSKeyedArchiver archivedDataWithRootObject:shortcut requiringSecureCoding:YES error:&error] forKey:SampleDefaultsKey];
+    XCTAssertNil(error);
     XCTAssertTrue([_monitor isShortcutRegistered:shortcut],
         @"Pass the shortcut from defaults to shortcut monitor after defaults change.");
 }
@@ -48,7 +58,9 @@ static NSString *const SampleDefaultsKey = @"sampleShortcut";
 {
     MASShortcut *shortcut = [MASShortcut shortcutWithKeyCode:1 modifierFlags:1];
     [_binder bindShortcutWithDefaultsKey:SampleDefaultsKey toAction:^{}];
-    [_defaults setObject:[NSKeyedArchiver archivedDataWithRootObject:shortcut] forKey:SampleDefaultsKey];
+    NSError *error;
+    [_defaults setObject:[NSKeyedArchiver archivedDataWithRootObject:shortcut requiringSecureCoding:YES error:&error] forKey:SampleDefaultsKey];
+    XCTAssertNil(error);
     [_defaults removeObjectForKey:SampleDefaultsKey];
     XCTAssertFalse([_monitor isShortcutRegistered:shortcut],
         @"Unregister shortcut from monitor after value is cleared from defaults.");
@@ -58,7 +70,9 @@ static NSString *const SampleDefaultsKey = @"sampleShortcut";
 {
     MASShortcut *shortcut = [MASShortcut shortcutWithKeyCode:1 modifierFlags:1];
     [_binder bindShortcutWithDefaultsKey:SampleDefaultsKey toAction:^{}];
-    [_defaults setObject:[NSKeyedArchiver archivedDataWithRootObject:shortcut] forKey:SampleDefaultsKey];
+    NSError *error;
+    [_defaults setObject:[NSKeyedArchiver archivedDataWithRootObject:shortcut requiringSecureCoding:YES error:&error] forKey:SampleDefaultsKey];
+    XCTAssertNil(error);
     [_binder breakBindingWithDefaultsKey:SampleDefaultsKey];
     XCTAssertFalse([_monitor isShortcutRegistered:shortcut],
         @"Unregister shortcut from monitor after binding was removed.");
@@ -67,7 +81,9 @@ static NSString *const SampleDefaultsKey = @"sampleShortcut";
 - (void) testRebinding
 {
     MASShortcut *shortcut = [MASShortcut shortcutWithKeyCode:1 modifierFlags:1];
-    [_defaults setObject:[NSKeyedArchiver archivedDataWithRootObject:shortcut] forKey:SampleDefaultsKey];
+    NSError *error;
+    [_defaults setObject:[NSKeyedArchiver archivedDataWithRootObject:shortcut requiringSecureCoding:YES error:&error] forKey:SampleDefaultsKey];
+    XCTAssertNil(error);
     [_binder bindShortcutWithDefaultsKey:SampleDefaultsKey toAction:^{}];
     [_binder breakBindingWithDefaultsKey:SampleDefaultsKey];
     [_binder bindShortcutWithDefaultsKey:SampleDefaultsKey toAction:^{}];